+++ src/features/Tag/model/useTag.ts
@@ -1,0 +1,13 @@
+import { atom, useAtom } from "jotai"
+import { Tag } from "../../../shared/type"
+
+const tagsAtom = atom<Tag[]>([])
+
+export const useTag = () => {
+  const [tags, setTags] = useAtom(tagsAtom)
+
+  return {
+    tags,
+    setTags,
+  }
+}
