+++ src/features/comment/model/useCommentDialog.ts
@@ -1,0 +1,16 @@
+import { atom, useAtom } from "jotai"
+
+const showAddCommentDialogAtom = atom<boolean>(false)
+const showEditCommentDialogAtom = atom<boolean>(false)
+
+export const useCommentDialog = () => {
+  const [showAddCommentDialog, setShowAddCommentDialog] = useAtom(showAddCommentDialogAtom)
+  const [showEditCommentDialog, setShowEditCommentDialog] = useAtom(showEditCommentDialogAtom)
+
+  return {
+    showAddCommentDialog,
+    setShowAddCommentDialog,
+    showEditCommentDialog,
+    setShowEditCommentDialog,
+  }
+}
