+++ src/widgets/comment/CommentItem.tsx
@@ -1,0 +1,73 @@
+import { Edit2, ThumbsUp, Trash2 } from "lucide-react"
+import { Button } from "../../shared/ui"
+import highlightText from "../../shared/ui/highlightText"
+import { useComment } from "../../features/comment/model/useComment"
+import { useCommentDialog } from "../../features/comment/model/useCommentDialog"
+import { useParam } from "../../shared/model/useParam"
+import { Comment } from "../../shared/type"
+
+export function CommentItem({ comment, postId }: { comment: Comment; postId: number }) {
+  const { comments, setComments, setSelectedComment } = useComment()
+  const { setShowEditCommentDialog } = useCommentDialog()
+  const { searchQuery } = useParam()
+
+  // 댓글 좋아요
+  const likeComment = async (id: number, postId: number) => {
+    try {
+      const response = await fetch(`/api/comments/${id}`, {
+        method: "PATCH",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({ likes: (comments[postId]?.find((c) => c.id === id)?.likes || 0) + 1 }),
+      })
+      const data = await response.json()
+      setComments((prev) => ({
+        ...prev,
+        [postId]: prev[postId].map((comment) => (comment.id === data.id ? data : comment)),
+      }))
+    } catch (error) {
+      console.error("댓글 좋아요 오류:", error)
+    }
+  }
+  // 댓글 삭제
+  const deleteComment = async (id: number, postId: number) => {
+    try {
+      await fetch(`/api/comments/${id}`, {
+        method: "DELETE",
+      })
+      setComments((prev) => ({
+        ...prev,
+        [postId]: prev[postId].filter((comment) => comment.id !== id),
+      }))
+    } catch (error) {
+      console.error("댓글 삭제 오류:", error)
+    }
+  }
+
+  return (
+    <div key={comment.id} className="flex items-center justify-between text-sm border-b pb-1">
+      <div className="flex items-center space-x-2 overflow-hidden">
+        <span className="font-medium truncate">{comment.user.username}:</span>
+        <span className="truncate">{highlightText(comment.body, searchQuery)}</span>
+      </div>
+      <div className="flex items-center space-x-1">
+        <Button variant="ghost" size="sm" onClick={() => likeComment(comment.id, postId)}>
+          <ThumbsUp className="w-3 h-3" />
+          <span className="ml-1 text-xs">{comment.likes}</span>
+        </Button>
+        <Button
+          variant="ghost"
+          size="sm"
+          onClick={() => {
+            setSelectedComment(comment)
+            setShowEditCommentDialog(true)
+          }}
+        >
+          <Edit2 className="w-3 h-3" />
+        </Button>
+        <Button variant="ghost" size="sm" onClick={() => deleteComment(comment.id, postId)}>
+          <Trash2 className="w-3 h-3" />
+        </Button>
+      </div>
+    </div>
+  )
+}
